# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.SubscriptionDTO:
    properties:
      end_date:
        type: string
      price:
        minimum: 0
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    required:
    - end_date
    - price
    - service_name
    - start_date
    - user_id
    type: object
  models.Subscription:
    properties:
      created_at:
        type: string
      endDate:
        type: string
      id:
        type: integer
      price:
        type: integer
      serviceName:
        type: string
      startDate:
        type: string
      updated_at:
        type: string
      userID:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API server for subscriptions
  title: Subscriptions API
  version: "1.0"
paths:
  /subscriptions:
    get:
      description: Calculating the total cost of subscriptions with filtering
      parameters:
      - description: User UUID
        in: query
        name: user_id
        type: string
      - description: Subscription service name
        in: query
        name: service_name
        type: string
      - description: Start period (MM-YYYY) or (DD-MM-YYYY)
        in: query
        name: from
        type: string
      - description: End period (MM-YYYY) or (DD-MM-YYYY)
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Calculating the total cost of subscriptions with filtering
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Creates a subscription
      parameters:
      - description: Subscription data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SubscriptionDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create subscription
      tags:
      - subscriptions
  /subscriptions/:
    get:
      description: Get all subscriptions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Subscription'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get all subscriptions
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      description: Delete subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Delete subscription
      tags:
      - subscriptions
    get:
      description: Get subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Subscription'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get subscription by ID
      tags:
      - subscriptions
    patch:
      consumes:
      - application/json
      description: Partially update subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SubscriptionDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Updated
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Partially update subscription
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Update subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated subscription data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SubscriptionDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Updated
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Update subscription
      tags:
      - subscriptions
swagger: "2.0"